+++ custom_spotting/__init__.py
@@ -2,7 +2,6 @@ from custom_spotting.actions import (
     Action,
     ACTION_CONFIGS,
     NUM_ACTION_CLASSES,
-    NUM_TEAM_ACTION_CLASSES,
     TRAINING_CE_RELATIVE_WEIGHTS,
     Team,
 )
@@ -20,7 +19,6 @@ __all__ = [
     "ACTION_CONFIGS",
     "TRAINING_CE_RELATIVE_WEIGHTS",
     "NUM_ACTION_CLASSES",
-    "NUM_TEAM_ACTION_CLASSES",
     "compute_map",
     "val_map",
     "infer_video",